requirements.txt
# ============================================================
# MLOps Pipeline for Customer Churn Prediction
# Requirements File
# ============================================================

# ---------- Core Data Science ----------
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scipy>=1.11.0

# ---------- ML Models ----------
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
imbalanced-learn>=0.11.0

# ---------- Hyperparameter Tuning ----------
optuna>=3.4.0

# ---------- Model Explainability ----------
shap>=0.43.0

# ---------- Experiment Tracking & Model Registry ----------
mlflow>=2.9.0

# ---------- Data Versioning ----------
dvc>=3.30.0
dvc-s3>=3.0.0          # Optional: for future S3 support

# ---------- Data Validation ----------
great-expectations>=0.18.0

# ---------- Workflow Orchestration ----------
prefect>=2.14.0

# ---------- API Serving ----------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6  # For file uploads (batch prediction)

# ---------- Monitoring & Metrics ----------
prometheus-client>=0.19.0
evidently>=0.4.0

# ---------- Interactive Demo ----------
streamlit>=1.29.0
plotly>=5.18.0
altair>=5.0.0

# ---------- Data Formats ----------
pyarrow>=14.0.0        # Parquet support
openpyxl>=3.1.0        # Excel support

# ---------- Configuration ----------
pyyaml>=6.0.0
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# ---------- Logging & Utilities ----------
loguru>=0.7.0
rich>=13.7.0
click>=8.1.0
httpx>=0.25.0          # Async HTTP client
tenacity>=8.2.0        # Retry logic

# ---------- Testing ----------
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
hypothesis>=6.91.0
httpx>=0.25.0          # For TestClient

# ---------- Code Quality ----------
ruff>=0.1.8
mypy>=1.7.0
pre-commit>=3.6.0

# ---------- Notebook Support ----------
jupyter>=1.0.0
ipykernel>=6.27.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ---------- Docker (for local builds) ----------
docker>=6.1.0
